Extract separate enviroments for Siri and GTFS
